refactor: move the common AllocGadget impls above int and uint
