+++ gadgets/src/traits/utilities/alloc.rs
@@ -14,6 +14,12 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
+use crate::utilities::{
+    boolean::{AllocatedBit, Boolean},
+    int::*,
+    integral::Integral,
+    uint::*,
+};
 use snarkvm_fields::Field;
 use snarkvm_r1cs::{errors::SynthesisError, ConstraintSystem};
 
@@ -137,3 +143,56 @@ impl<I, F: Field, A: AllocGadget<I, F>> AllocGadget<[I], F> for Vec<A> {
         Ok(vec)
     }
 }
+
+macro_rules! alloc_gadget_fn_impl {
+    ($gadget: ident, $fn_name: ident) => {
+        fn $fn_name<
+            Fn: FnOnce() -> Result<T, SynthesisError>,
+            T: Borrow<<$gadget as Integral>::IntegerType>,
+            CS: ConstraintSystem<F>,
+        >(
+            mut cs: CS,
+            value_gen: Fn,
+        ) -> Result<Self, SynthesisError> {
+            let value = value_gen().map(|val| *val.borrow());
+            let values = match value {
+                Ok(mut val) => {
+                    let mut v = Vec::with_capacity(<$gadget as Integral>::SIZE);
+
+                    for _ in 0..<$gadget as Integral>::SIZE {
+                        v.push(Some(val & 1 == 1));
+                        val >>= 1;
+                    }
+
+                    v
+                }
+                _ => vec![None; <$gadget as Integral>::SIZE],
+            };
+
+            let bits = values
+                .into_iter()
+                .enumerate()
+                .map(|(i, v)| {
+                    Ok(Boolean::from(AllocatedBit::$fn_name(
+                        &mut cs.ns(|| format!("allocated bit_gadget {}", i)),
+                        || v.ok_or(SynthesisError::AssignmentMissing),
+                    )?))
+                })
+                .collect::<Result<Vec<_>, SynthesisError>>()?;
+
+            Ok(Self::new(bits, value.ok()))
+        }
+    };
+}
+
+macro_rules! alloc_gadget_int_impl {
+    ($($gadget: ident)*) => ($(
+        impl<F: Field> AllocGadget<<$gadget as Integral>::IntegerType, F> for $gadget {
+            alloc_gadget_fn_impl!($gadget, alloc);
+
+            alloc_gadget_fn_impl!($gadget, alloc_input);
+        }
+    )*)
+}
+
+alloc_gadget_int_impl!(Int8 Int16 Int32 Int64 Int128 UInt8 UInt16 UInt32 UInt64 UInt128);
