+++ gadgets/src/traits/utilities/int/utilities/alloc.rs
@@ -16,74 +16,10 @@
 
 use crate::{
     fields::FpGadget,
-    utilities::{
-        alloc::AllocGadget,
-        boolean::{AllocatedBit, Boolean},
-        eq::EqGadget,
-        int::*,
-        integral::Integral,
-        ToBitsBEGadget,
-    },
+    utilities::{alloc::AllocGadget, boolean::Boolean, eq::EqGadget, int::*, integral::Integral, ToBitsBEGadget},
 };
-use snarkvm_fields::{traits::to_field_vec::ToConstraintField, Field, FpParameters, PrimeField};
+use snarkvm_fields::{traits::to_field_vec::ToConstraintField, FpParameters, PrimeField};
 use snarkvm_r1cs::{errors::SynthesisError, ConstraintSystem};
-
-use core::borrow::Borrow;
-
-macro_rules! alloc_int_fn_impl {
-    ($gadget: ident, $fn_name: ident) => {
-        fn $fn_name<
-            Fn: FnOnce() -> Result<T, SynthesisError>,
-            T: Borrow<<$gadget as Integral>::IntegerType>,
-            CS: ConstraintSystem<F>,
-        >(
-            mut cs: CS,
-            value_gen: Fn,
-        ) -> Result<Self, SynthesisError> {
-            let value = value_gen().map(|val| *val.borrow());
-            let values = match value {
-                Ok(mut val) => {
-                    let mut v = Vec::with_capacity(<$gadget as Integral>::SIZE);
-
-                    for _ in 0..<$gadget as Integral>::SIZE {
-                        v.push(Some(val & 1 == 1));
-                        val >>= 1;
-                    }
-
-                    v
-                }
-                _ => vec![None; <$gadget as Integral>::SIZE],
-            };
-
-            let bits = values
-                .into_iter()
-                .enumerate()
-                .map(|(i, v)| {
-                    Ok(Boolean::from(AllocatedBit::$fn_name(
-                        &mut cs.ns(|| format!("allocated bit_gadget {}", i)),
-                        || v.ok_or(SynthesisError::AssignmentMissing),
-                    )?))
-                })
-                .collect::<Result<Vec<_>, SynthesisError>>()?;
-
-            Ok(Self {
-                bits,
-                value: value.ok(),
-            })
-        }
-    };
-}
-
-macro_rules! alloc_int_impl {
-    ($($gadget: ident)*) => ($(
-        impl<F: Field> AllocGadget<<$gadget as Integral>::IntegerType, F> for $gadget {
-            alloc_int_fn_impl!($gadget, alloc);
-            alloc_int_fn_impl!($gadget, alloc_input);
-        }
-    )*)
-}
-
-alloc_int_impl!(Int8 Int16 Int32 Int64 Int128);
 
 /// Alloc the unsigned integer through field elements rather purely bits
 /// to reduce the number of input allocations.
